+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Subsystems/IntakeSubsystem.java
@@ -3,6 +3,7 @@ package org.firstinspires.ftc.teamcode.Subsystems;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
+import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.teamcode.Utilities.Constants;
 
 import java.util.concurrent.TimeUnit;
@@ -12,10 +13,17 @@ public class IntakeSubsystem {
     private double intakePower;
 
     private ElapsedTime runtime;
+    Telemetry telemetry;
 
-    public IntakeSubsystem(DcMotor intake, ElapsedTime runtime) {
+    public enum Modes {
+        INTAKE,
+        OUTTAKE
+    }
+
+    public IntakeSubsystem(DcMotor intake, ElapsedTime runtime, Telemetry telemetry) {
         this.intake = intake;
         this.runtime = runtime;
+        this.telemetry = telemetry;
         initialize();
     }
 
@@ -27,6 +35,18 @@ public class IntakeSubsystem {
     public void teleOPIntake(double gamepad2RightTrigger, double gamepad2LeftTrigger) {
         intakePower = gamepad2RightTrigger - gamepad2LeftTrigger;
         intake.setPower(intakePower);
+    }
+
+    public void intake() {
+        intake.setPower(Constants.autoIntakePower);
+    }
+
+    public void outtake() {
+        intake.setPower(-Constants.autoIntakePower);
+    }
+
+    public void stopMotor() {
+        intake.setPower(0.0);
     }
 
     public void autoIntake(Modes mode, double timeSeconds) {
@@ -46,10 +66,5 @@ public class IntakeSubsystem {
 
         intakePower = 0.0;
         intake.setPower(intakePower);
-    }
-
-    public enum Modes {
-        INTAKE,
-        OUTTAKE
     }
 }
