+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Tests/PositionTest.java
@@ -27,8 +27,9 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode.OpModes;
+package org.firstinspires.ftc.teamcode.Tests;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
@@ -38,15 +39,17 @@ import com.qualcomm.robotcore.hardware.TouchSensor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
 import org.firstinspires.ftc.teamcode.Subsystems.ArmSubsystem;
+import org.firstinspires.ftc.teamcode.Subsystems.DrivetrainSubsystem;
 import org.firstinspires.ftc.teamcode.Subsystems.ExtenderSubsystem;
 import org.firstinspires.ftc.teamcode.Subsystems.IntakeSubsystem;
 import org.firstinspires.ftc.teamcode.Subsystems.WristSubsystem;
-import org.firstinspires.ftc.teamcode.Subsystems.DrivetrainSubsystem;
 import org.firstinspires.ftc.teamcode.Utilities.Constants;
+import org.firstinspires.ftc.teamcode.Utilities.Templates.CommandTemplate;
+import org.firstinspires.ftc.teamcode.Utilities.Templates.SubsystemTemplate;
 
-@TeleOp(name="Main TeleOp", group="Linear OpMode")
+@TeleOp(name="Position Test", group="Linear OpMode")
 //@Disabled
-public class MainTeleOp extends LinearOpMode {
+public class PositionTest extends LinearOpMode {
 
     // Declare OpMode members.
     private ElapsedTime runtime = new ElapsedTime();
@@ -89,30 +92,16 @@ public class MainTeleOp extends LinearOpMode {
         // run until the end of the match (driver presses STOP)
         while (opModeIsActive()) {
 
-            drivetrainSubsystem.driveManual(gamepad1.left_stick_x, gamepad1.left_stick_y, gamepad1.right_stick_x);
-            intakeSubsystem.teleOPIntake(gamepad2.right_trigger, gamepad2.left_trigger);
-            wristSubsystem.manualAngleWrist(gamepad2.right_bumper, gamepad2.left_bumper);
-            armSubsystem.ManualPositionArm(gamepad2.dpad_left, gamepad2.dpad_right);
+            int[] driveCounts = drivetrainSubsystem.getDriveMotorCounts();
 
-            if (gamepad2.dpad_up) {
-                wristSubsystem.autoAngleWrist(WristSubsystem.Positions.UP);
-            }
-            else if (gamepad2.dpad_down) {
-                wristSubsystem.autoAngleWrist(WristSubsystem.Positions.DOWN);
-            }
-
-            if (gamepad2.x) {
-                extenderSubsystem.extendToPosition(ExtenderSubsystem.Position.RETRACTED);
-            }
-            else if (gamepad2.y) {
-                extenderSubsystem.extendToPosition(ExtenderSubsystem.Position.HALFWAY);
-            }
-            else if (gamepad2.b) {
-                extenderSubsystem.extendToPosition(ExtenderSubsystem.Position.EXTENDED);
-            }
-
-            // Show the elapsed game time and wheel power.
+            // Show the elapsed game time
             telemetry.addData("Status", "Run Time: " + runtime.toString());
+            telemetry.addData("Drive Motors",
+                    "BackLeft, BackRight, FrontLeft, FrontRight",
+                    driveCounts[0], driveCounts[1], driveCounts[2], driveCounts[3]);
+            telemetry.addData("Arm", armSubsystem.getArmPosition());
+            telemetry.addData("Extender", extenderSubsystem.getExtenderPosition());
+            telemetry.addData("Wrist", wristSubsystem.getWristPosition());
             telemetry.update();
         }
     }
